requirements.txt
Django==4.0.3
psycopg2-binary==2.9.3
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6

